api処理の関数を責務毎に分割
