+++ features/teams/components/InviteTeamForm.tsx
@@ -3,13 +3,13 @@ import { getSession } from '@/features/sessions/functions/session';
 import { useRouter } from 'next/router';
 import { FormContext } from '@/context/FormContext';
 import { InitialInviteTeamFormData, resetFormValue } from '@/common/functions/form';
-import { callInviteTeam } from '@/common/functions/api';
 import { getApiHeaders } from '@/common/functions/api';
 import { dataConfirmAlert } from '@/common/functions/form';
 import FormSubmitButton from '@/components/ui-elements/FormSubmitButton';
 import FormCancelButton from '@/components/ui-elements/FormCancelButton';
 import InviteTeamFormGroup from './InviteTeamFormGroup';
 import { InviteTeamFormContext } from '../contexts/InviteTeamFormContext';
+import { callInviteTeam } from '../functions/api';
 
 const InviteTeamForm: FC = () => {
   const { inviteTeamFormData, setInviteTeamFormData } = useContext(InviteTeamFormContext);
